Implement `CmdStreamHandler` for streaming command and a dummy command using AsyncPipe to validate behavior
Summary:
### Context
Add capability of supporting streaming commands which can output large amounts of data at run-time without effecting the performance of the commands. We will need changes at the framework level on how we deal with command handlers.

`CmdStreamHandler` is implemented to handle streaming related operations. A dummy streaming command is also created to imitate the behavior of a normal streaming command.

### Changes
Implement unittest to verify the implementation of a dummy stream command
- Subclass `CmdHandler` to be `CmdStreamHandler` and override the run method
- Implement a dummy command using `AsyncPipe` to validate behavior
- `queryClient` in the `StreamCmdHandler` now returns an generator, handle this using coroutine features and print when receiving the data

### Next Steps
- Add unittest
- Pull out as much code into common super
- Make sure term signals are handled gracefully

Reviewed By: peygar

Differential Revision: D36866722

fbshipit-source-id: 110d27bd4e48f3155fb1cb4d8841ed8f0860a38b