Allow skipping coldboot in initExternalPhyMap
Summary:
The new xphy programming logic in qsfp_service always tries to use the warmboot/coldboot files when determining how to initialize the phys.

When we run this code from fboss-xphy, the warmboot state file won't exist so we'll default to coldbooting and reset all the phys on the system.

To fix these, we'll just add a parameter to skip checking the warmboot file.

Reviewed By: rajank7

Differential Revision: D37008691

fbshipit-source-id: ebdd0a1bc50c98ae37b976126afedc807ca2b269