Fix the order in which moduleCapabilities_ is updated
Summary:
This diff fixes a heap-use-after-free TSAN violation that I saw while bringing up MPH2.

The reason is that in updateCachedTransceiverInfoLocked() we first read numMediaLanes which can return a default value of 4 if the capabilities hasn't been updated yet.

```
unsigned int CmisModule::numMediaLanes() const {
  auto application = static_cast<uint8_t>(getSmfMediaInterface());
  auto capabilityIter = moduleCapabilities_.find(application);
  if (capabilityIter == moduleCapabilities_.end()) {
    return 4;
  }
  return capabilityIter->second.mediaLaneCount;
}
```

Based on the count 4, we create 4 channels in TransceiverInfo and call getSensorsPerChanInfo() to update the channel information. But before the call to getSensorsPerChanInfo(), we call getTransceiverSettingsInfo() which calls getApplicationCapabilities() and updates the moduleCapabilities_.

Now inside getSensorsPerChanInfo(), there is another call to numMediaLanes() which this time can return a value different than 4. It was 8 in my case.

We only allocated space for 4 channels earlier, but try to access 8 channels in getSensorsPerChanInfo() causing the heap-use-after-free violation.

The fix is to move getApplicationCapabilities to updateQsfpData so that it's updated every time the eeprom is read. Anything that happens after updateQsfpData (like updateCachedTransceiverInfoLocked) will always see consistent values from numMediaLanes().

Differential Revision: D37077887

fbshipit-source-id: 88395b6b7cc2cbf0e4ad2a908dd7f8b6163bb237