Implement valid filters for the command show port and introduce filter validation during run()
Summary:
Include code for filter validation that checks that the keys and values passed from command line are as expected. It also checks for the data type of the value and the validity of the filter operation.

Regarding getValidFilters() method in CmdShowPort(): We had two options:
1) Keep the valid filters map as a member field in the class
2) have a method that creates this map on the fly and returns.

We decided to go with the second approach because this method will only be called once and also in case when no filter arguments have been passed from the command line, it will avoid creating unnecessary instances of this map.

Reviewed By: zechengh09

Differential Revision: D37084565

fbshipit-source-id: 2eb55d5d8567010824f987c3eab78dd554b4276e