Implement help command unit tests
Summary:
Implement unit tests to that extensively check the output of the help command. This test does 100% test coverage by iterating over all the commands (and all their recursively nested subcommands) in the command tree to verify the correct help output is generated for all the nested levels!

The tree traversal has been implemented in an iterative in-order manner.

Reviewed By: peygar

Differential Revision: D37079368

fbshipit-source-id: 1220d34a4bb5d549dfc27dc72b9cf4aac8167a21