Fix l2 learning callback handles for bcm devices
Summary:
See https://fb.workplace.com/groups/280091245787228/permalink/1392776621185346/ for more details.

There's no real hw learning on TH3/TH4, hence not needing to validate/remove the L2 entry in hardware when the entry is learned/aged. We should only program those entry only when the static flag or trunk member flag is set.

To solve this - read from the SDK to see the current l2 entry programmed in the hardware. If there's any discrepancy, add the updated l2 entry to the SDK (which would internally do a remove and create).

Reviewed By: shri-khare

Differential Revision: D36921021

fbshipit-source-id: 559a524fc9ffbb63eaab0c46f99b42280493dec2