Implement and verify `doTyping` for argument type conversion
Summary:
### Context
Currently all fboss2 subcommands accept arguments as a string vector and the framework does not do any argument parsing/validation automatically. We can use the enum type of the argument to do some of this at the framework level and pass a fully typed struct to the command handler.

Based upon the single instance type conversion, we hope to support the type conversion with templates so that the implementation will not be restricted by various forms of the actual input

### Changes
- Implement a tuple version `DoTyping` and verify with test cases
- Implement and verify the template version of `DoTyping`

### Next Steps
- Create argument type class
- Replace them in actual implementation

Reviewed By: peygar

Differential Revision: D37228864

fbshipit-source-id: a7e9773aad3b8f82f9c9ab66f4e9d8734f4b685a