Implement `doTyping`, add Arg traits, and validate `ProbState`
Summary:
### Context
Implement `doTyping` as the last step of `getTypedArgs()` for type conversion

Because we are using the new typed class to replace the previous untyped `std::vector<std::string>` as container to store arguments. Directly inheriting from `std::vector` is a dangerous behavior, so we decided to implement some of its traits.

Add validation to arg type `ProbState`.

Reviewed By: peygar

Differential Revision: D37329433

fbshipit-source-id: 363016a787adf5ef4bef7a05e1f5fdff6cd9d1ef