Move implementation of GetValidFiltersGeneric from facebook/CmdHandler.cpp to CmdHandler.cpp and fix OSS linking errors
Summary: same as title. Earlier organization (Implementation inside Facebook/CmdHandler.cpp) was not really needed. Originally, it was added to the facebook/ directory because we thought that thrift reflections would not work with OSS. But we now know that the OSS linking errors were not because of this reason. Hence, these implementations have been moved back to CmdHandler.cpp (outside facebook/).

Reviewed By: zechengh09

Differential Revision: D37433045

fbshipit-source-id: 48a30f0d62f0a0f401bcc6b4273f8bc168002610