Introduce component syncers
Summary:
Introducing a concept of component syncers which will be used to manage the publishing of a single component. For examples, agent components are switch state, config, switch stats and lldp.

My main goals here are to
- provide a clearer interface to the pub sub model (requiring an intial sync)
- abstract out some of the event base and state management internals of pub sub manager / stream client
- modularize component publishing so that each component can own its own respective syncer rather than a service wide syncer. i.e. LldpManager can own a FsdbLldpStateSyncer rather than requiring a reference to the general agent's FsdbSyncer.

The third point became a headache while I was working on lldp publish. since we'd either have to have lldp depend on FsdbSyncer (bad since it also includes switch state and other components), or have to thread updates through SwSwitch which I didn't like. I imagine this problem will come up with other services that have a more decentralized state too.

I'm also thinking of this as a general way to eventually hook up a state managed by CowStateMgr to fsdb publishing

Reviewed By: zechengh09

Differential Revision: D37310969

fbshipit-source-id: b81a682daf06111fdb4d6c59611c4a4e9162443f