Enable filtering on other show commands
Summary:
To make sure that filtering does not break other parts of the code as people add more commands, we only enable filtering only few commands. For now, it's enabled on show arp, show ndp, show aggregate-port and show port.

Extending to others is easy - just add a constexpr flag into the command's command handler and set it to true!

Reviewed By: zechengh09, msomasundaran

Differential Revision: D37726278

fbshipit-source-id: e4525e550d0365e432a10a60db851ec5f8d5c20e