rackmon2: Fix failed probe on wedge100
Summary:
A couple of assumptions were wrong.

First and foremost, the IOCTL TIOCSERGETLSR is non-blocking.
Thus, we cannot assume that we do not need to do a second
iteration on it (The comment was wrong).
Hence, as observed on wedge100, we end up with a lot of
prints "Waited loops: 100". Change this to print only
when we hit the max iterations, not every time we have
more than 1 iteration.

Secondly, 100 iterations is very often exceeded. We would need
to bump this up to 10k to bring in some reliability.

Imported from: D36840307

Reviewed By: tao-ren

Differential Revision: D37766469

fbshipit-source-id: d51f02cec6cce0c78dad235d86a9ed46ed373f7b