rackmon2: Move profile information to logging output
Summary:
This essentially moves the profile information to use
glog's output to allow it to be streamed out just like
we would stream out logs.

This means we can remove all the complicated means of
holding the profiling data as stringstream and have
an API to retrieve them.

In addition, we can now profile interaction spanning
multiple processes. Example, I can have profiling
turned on in rackmoncli to measure the IPC as well
and do not need to figure out how to interleave the
profile data.

Imported from: D36424853

Reviewed By: tao-ren

Differential Revision: D37766470

fbshipit-source-id: a4c2575240abe765f0663bf96a8f9e04f24e2d70