ensure that entries are always created in the order of their keys for node map
Summary:
items() call return entries in the order they're found in F14NodeMap which is mainly an unordered map. The order in which these entries are returned, was observed to be different across different runs. This leads to inconsistency in tests which verify exact dynamic objects, since the different order of items leads to different "array" in dynamic object.

this diff enforces the order in NodeMap's entries array, sorted by keys of items.

Reviewed By: zechengh09

Differential Revision: D37790481

fbshipit-source-id: bfddaec16913a3068c56002a8e635a72fe0604c9