Replace operator enum with the operator class in getFilterOp and related places. Also update the tests to work with this new design
Summary:
Earlier, we had an enum for the operators which is now being replaced by a class for each operator that has its compare method (previous diff).
In this diff, I integrate those operator classes into the code and replace the usage of the operator enum with the corresponding operator class.

Also update the tests to work with this change

Reviewed By: zechengh09

Differential Revision: D37694089

fbshipit-source-id: 38b88dd88af8af9c7cf1ebf253c809cc1a3cbc70