Infer the media type from interfaceType for SAI implementations
Summary:
As we found in S280146, the media type is configured as SAI_PORT_MEDIA_TYPE.OPTICAL on W400 uplinks instead of SAI_PORT_MEDIA_TYPE.BACKPLANE as is configured with native sdk implementations. This causes port flaps over time.

This is happening because we are currently inferring the media type from the 'medium' field in PlatformMapping for SAI but we infer it from 'interfaceType' for non-SAI. This change introduces a flag (so that we can stage disruptive upgrades) that allows us to infer the media type correctly (and same as native impl) for SAI

Reviewed By: peygar, zechengh09

Differential Revision: D37810015

fbshipit-source-id: b69541856821a315039be8990bcc8315e98c463a