Simplify getIntfAddrToReach -> getIntfToReach
Summary: Simplifying this function to just return the node rather than this one off struct. If anyone needs the actual ip addr we can get it from the interface. Didn't seem good to return a pointer to the intf and a pointer to the ip addr which was a key in a map anyways.

Reviewed By: zechengh09

Differential Revision: D37872756

fbshipit-source-id: 823f26e83d1d2546529fcbcd8ba996c6a820c676