Add PFC mode and priority attributes to SAI port
Summary:
Add the 2 PFC port attributes to specify the PFC mode and the priorities to enable PFC on.
With the current support available, PFC mode can be set to COMBINED indicating PFC
tx==rx, but as we have different configs for rx and tx, we'll need to include support for
mode SEPARATE eventually. Until then, will just use the attribute
SAI_PORT_ATTR_PRIORITY_FLOW_CONTROL to specify the rx/tx priorities.

Reviewed By: shri-khare

Differential Revision: D37778782

fbshipit-source-id: 6981c0e9cd9234913aa2200dacfd01fc96c36928