Fix for crash in lacp distruptive test
Summary:
In multinode lacp test, we have a disruptive test where the lacp state machines are stopped. As part of this, the subport is erased from a internal map. If there are other events such as a link flap during this time, the map gets accessed and asserts since subport is not found.

Fix: don't remove the subport from map. this api is used only in test. So no impact to prod

Reviewed By: peygar

Differential Revision: D37999016

fbshipit-source-id: 5c27444a6041b9d9f976a348cd5eb8dcc8841994