Improved filter parsing
Summary:
A small, obvious, pending improvement in the filter parsing. Earlier, we required input filter strings to be of the form TERM1&&TERM2||TERM3||TERM4... where each TERMi is of the form <key op value>.

Here, the restriction was that there can be no spaces between the logical operator and the TERM string - which is quite restrictive. This diff introduces a simple fix to get more flexibility in inputting filter predicates.

Now, one can enter a filter term of the form  `linkState == Up && id <= 5`, which is more intuitive and easy. (The change is that spaces between `Up` and `&&`, `&&` and `id` are now acceptable)!

Also wrote a new unit test for this change.

Reviewed By: zechengh09

Differential Revision: D37888946

fbshipit-source-id: 5bfed4f731c0c6d9b6b0d61b113926c27facc0ef