Fix heap-use-after-free in reading eye values
Summary:
Currently the API to read eye values returns a vector of sai_port_lane_eye_values_t which we then try to convert to sai_port_eye_values_list_t. We return the pointer to the underlying data in the vector from getPortEyeValues function and use it later. However, the vector gets cleaned up at the end of getPortEyeValues causing a asan warning of "heap-use-after-free" in the caller of getPortEyeValues.
I noticed that when running an asan build link of test in the stacked diff.

This diff fixes it by directly returning the vector instead of converting it to sai_port_eye_values_list_t type.

Differential Revision: D38026470

fbshipit-source-id: dacd0e58aebe49e2b982b9461f2234ca745ae1aa