Added support for SAI_PORT_ATTR_ERR_STATUS_LIST SAI attribute
Summary:
This attribute provides a list of error status. We are mainly interested to know if the following errors exist - SAI_PORT_ERR_STATUS_SIGNAL_LOCAL_ERROR, SAI_PORT_ERR_STATUS_REMOTE_FAULT_STATUS. These are essentially the local fault status and remote fault status that will be added to the PhyInfo collection in the next diff.

Defined a HwAsic feature for this diagnostic since not all ASICs will support this

Reviewed By: jasmeetbagga

Differential Revision: D38041334

fbshipit-source-id: ba8ce18634155c4cb9988b59670c399e458b842d