Thrifty map deep legacy conversions
Summary:
The current legacy conversions for node map only did legacy for the map itself but used the thrifty conversions for the actual nodes. This was fine for the most part in tests because we were manually verifying each node as well as the map conversion. But need to fix this if we want to revert a serilzation back to legacy (for fibs)

This actually did surface a bug in sflow conversion due to the primary key missing in the legacy conversion. Had to do some gymnastics to quickly fix this

Reviewed By: jasmeetbagga

Differential Revision: D37969902

fbshipit-source-id: 9ace834edc369f9ee6ecd8bfe06da52acb6fc513