Initial setup and implement first test case `PfcDefaultProgramming`
Summary:
### Context
SAI is Software Abstract Interface which is standard APIs used across different platforms such as BCM, CSCO. We want to refactor all Bcm PFC tests in BcmPfcTests.cpp so that they are ready to be used for SAI as well.
At the end of the task. (1) BCM tests should build, (2) run as-it-is from new location. (as mentioned below)

We will stub out the SAI functionality to be added later, so these tests DO NOT need to run on SAI as part of this task. We also need to remove these tests from BcmPfcTests.cpp so we do NOT have duplicates.

### Changes
- Created `HwAgentPfcTests.cpp` and implement the first test case `TEST_F(BcmPfcTests, PfcDefaultProgramming)`
- Created header files and corresponding implementations in different files.

Differential Revision: D37951173

fbshipit-source-id: fac776cc53d859cbd4c19da43c2584aab8014921