Fixing port flap caused by a qsfp service warm boot after an agent cold boot.
Summary:
The following sequence is currently broken after the introduction of new port programming
1. cold boot agent + qsfp
2. warm boot qsfp ---------> This causes port flaps

This is because of the following code https://fburl.com/code/fxwph3cy
When qsfp service warm boots in step #2, it retrieves the last cold boot time from agent which is non-zero. QSFP service thinks it needs to cold boot as well and resets the data path in optics causing ports to flap.

In this change, I update the agent config timestamps in the warm boot state file that is written on a graceful exit. On a qsfp service warmboot, the qsfp service will retrieve the timestamps from this cache and use that to figure out if it needs to cold boot or warmboot.

Reviewed By: rajank7

Differential Revision: D38132129

fbshipit-source-id: 8e7fefdb1bd46f69b2c2832471cc45b477751526