provide accessor methods in route prefix
Summary:
as titled. this is basically intended to start using thrift storage for this object.

Main change is in RouteTypes.h and RouteTypes.cpp.

All the users of RoutePrefix were directly access network and mask fields of RoutePrefix. The usage is replaced with corresponding calls to access these fields. A constructor is used wherever these fields were set.

Reviewed By: jasmeetbagga

Differential Revision: D38011048

fbshipit-source-id: 45401d281c7c7563255721e33290e4eecc536e36