rackmonsvc: Add support for filtered monitor data
Summary:
This allows someone to filter data to return only the required values.
The user is allowed to apply device filters (return all or subset of devices (filtered by address or type) or/and
register filters (return all or subset of registers (filtered by address or name).

This also adds the capability to return only the latest value if the monitored register stores a history of values.

Reviewed By: jasmeetbagga

Differential Revision: D38141686

fbshipit-source-id: 094c589c26702035c0e6031fdd5dac3dc1c501c5