Handle different PrbsPolynomial on A and Z components in PrbsTest
Summary:
When enabling PRBS on a ASIC <-> Transceiver_System link, the prbs polynomial could be different on both components. For PAM4 speeds, optics generally use PRBS31Q (gray encoded) polynomial but the ASIC/XPHY still require us to configure PRBS31 (internally they gray encode it). The current implementation of PrbsTest assumed we'll always use the same PRBS Polynomial on both ends. Therefore, this change refactors the test so that we can specify different polynomials on either end of the link.

In addition to interface name and component, we now have a third property PrbsPolynomial to store per interface and I do that using a new TestPort struct.

Differential Revision: D37935946

fbshipit-source-id: 4f2ff776d7cf3b97c982eb03b56ccc217bf900cc