SAI log setting MAX API check correction
Summary:
The PHY initialization for the Sandia platform failed with the message:

api type invalid 36

This happens because for Sandia the SAI SDK has SAI_API_MAX value as 45. But the switch case in saiApiTypeToString() throws error if api type is more than 35.

To fix this we are returning  "unsupported" for api type 36-45 and throw error only if api type is more than that.

Differential Revision: D38345928

fbshipit-source-id: 4db25bc48836937040870de1bafd9ea01eb26161