Using ThreadHeartbeat to detect stuck state machine threads in qsfp_service
Summary:
In this diff, I use the ThreadHeartbeat class to schedule a timeout every 10 seconds on the per module updateStateMachine thread and the global state machine thread. Then, we use the ThreadHeartbeatWatchdog to monitor these heartbeats and increment a counter when the watchdog doesn't detect a heartbeat for 10 x heartbeat_interval on any thread.

In the next diff, I'll start publishing the counter to ODS

Differential Revision: D38551010

fbshipit-source-id: a63444bc6681c4a4997e7b55b32f0f3849c8809f