Add a thrift API to support transceiver hard reset
Summary:
Currently, "wedge_qsfp_util --qsfp-hard-reset" bypasses qsfp_service and resets transceivers directly even when the –direct-i2c option is not specified. During a hard reset, a transceiver's reset line will be asserted. The transceiver will be unresponsive, and its link will go down. When its reset line is released, it will need to be reprogrammed. However, because the –qsfp-hard-reset option bypasses qsfp_service, the transceiver state machine is not aware of this reset event and thus, does not reprogram the transceiver after the reset.

A solution is to force the –qsfp-hard-reset option to go through qsfp_service via a thrift call. When qsfp_service is not running or the –direct-i2c option is used with the –qsfp-hard-reset option, the command will fail. Otherwise, qsfp_service will hard reset the select transceivers and transition their state machine to their NOT_PRESENT state.

This diff adds a thrift API to allow wedge_qsfp_util to request qsfp_service to hard reset a select transceiver.

Reviewed By: rajank7

Differential Revision: D38557121

fbshipit-source-id: 1aedb4c89ccf83341789c151354ddc27fd920365