Move HwTest cases to use WITH_RETIRES toolkit
Summary:
I introduced this tool kit a while back as a replacement for checkWithRetries in tests to improve error messages when tests fail. For example in resetTranscieverAndDetectStateChanged we'll now get the exact check and line number of failure, along with the optional log.

Can probably be smarter in HwStateMachineTest to get better idea of whats actually failing but for now just trying to convert usages of checkWithRetry to reduce confusion

Reviewed By: birdsoup

Differential Revision: D38362587

fbshipit-source-id: 911f43fd71216f241f29fcf9b160a5a18714a90a