Added debug to print lane_list for creating SAI PHY port
Summary:
The Sandia SDK does not support "get attribute" function for HW_LANE_LIST so for debugging purpose we need to dump the lane list from the PHY port create function.

Also we don't want to print each attributes separately because the ports get created in parallel so the output will be mixed together. So we print to a temporary string and then output the temp string.

Reviewed By: harshitgulati18

Differential Revision: D38695327

fbshipit-source-id: 861ca63fa51481b9d275696c7ab5bc09f4e767c6