Allow toggling a GPIO output line
Summary: Fixed a bug in the GpiodLine::setValue() function to allow its callers to toggle a GPIO output line, e.g., reset then unreset a transceiver. Previously, once a GPIO output line is reservered in the first call to the GpiodLine::setValue() function, the second call to this function would fail because it coud not reserve this GPIO output line anymore. The fix is to add a bool reserveLine param to this function to let its callers decide whether or not they need to reserve a GPIO output line before manipulating it.

Reviewed By: rajank7

Differential Revision: D38819144

fbshipit-source-id: aa1fa339a4fac24a62f9344050cbcbbbd4914a50