Set uninit inDiscards*, inPause counters to 0
Summary:
These counters are not available for fabric ports,
while for others they will just be collected in updateStats
call. So set them to 0, to keep CurAndPrev counter utils
happy. Alternatively,
- I could have set these to 0 only if they were not in supported stats.
But since supported sets is a vector, this is a O(n) lookup
OR
- Set to 0 for only fabric ports. This assumes that the counters
would never be available for fabric ports and introduces a unnecessary
coupling.
So just always set them to 0 and rely on collection to fill them up

Reviewed By: nivinl

Differential Revision: D38912332

fbshipit-source-id: f9b68b2804241bb5af47dee9ae7baf48344d392c