Fix resetDataPathLog
Summary:
We should only log that we're resetting the datapath when resetDataPath is true.

Without this change, we'd always (incorrectly) log that we're resetting the data path if both configs specify a last coldboot config timestamp, even if the timestamps are the same. This makes debugging confusing since it looks like we're going to reset all of the ports, but we don't actually do so.

Reviewed By: harshitgulati18

Differential Revision: D38843473

fbshipit-source-id: c1a39da2a0b5f68c592fe10773de6793f4cc75ea