Correctly cache the number of IPHY lanes for a BcmPort
Summary:
Currently we use getIphyLaneConfigs to find the number of phy lanes that a port uses. However, getIphyLaneConfigs only returns the lanes that need to override the tx settings which is <= total_lanes. Sometimes (for example some ports on wedge100s) don't override tx settings, hence the number of phy lanes on those ports is currently returned 0.

This diff fixes that bug by using the number of pins (defined in PlatformMapping) as the number of phy lanes.

Reviewed By: zechengh09

Differential Revision: D39092872

fbshipit-source-id: 3d7ac47a6334be0c4cbcd853e4dfccfc6cd87e50