Add option to specify router address to use in RA via config
Summary:
One of the causes of S289408 was that we had a mismatch
in GW addresses b/w V4 and V6 on servers. V4 just uses fe80::face:b00c
while v6 used the interface link local. Now in absence of v4 traffic,
v4 GW was not resolved. On top of it, SHIV was asserting for both
GWs to be resolved before starting up, else it would crash and exit.
This check was clearly wrong (since SHIV would advertise the V4 VIPs
which would trigger V4 traffic) and has since been fixed. However
its best of us to unify the GW addresses.

Reviewed By: peygar

Differential Revision: D39089872

fbshipit-source-id: aeb3ae13f7e04bc6b99d1a12c527734389a738bf