Keeping share pointer to aclTableMap to avoid generating different objects
Summary:
After flattening thrift struct to inline map objects, invoking `getAclTableMap()` would construct a new object (though the returned AclTableMap is the same). This would break delta comparison on share pointers.

To avoid this, we still keep a reference to the share pointer in AclTableGroupFields. The pointer and thrift struct is updated simultaneously to avoid the issue.

Reviewed By: peygar

Differential Revision: D39110295

fbshipit-source-id: bac827f9225c7f5f9811ae2e1bb76338d35f930c