Allow for hitless rollback of RA.routerAddress config
Summary:
Previous diffs added the capability of specifying src IP to use
in router advertisements. When specified, we would use well known addresses
(say fe80::face:b00c) to unify the GW address across servers. Previously
this was non uniform as the rack switch's link local was used in RA adv.
This is all fine and good, we we also need to consider hitless rollback
of this config. In such a scenario, GW will again become the RSW link
local address. Now we may find that the RSW LL address is no longer
resolved at the server. In which case we would see a few packets
of loss as servers resolve this address. To mitigate that, when
RA.routerAddress is specified, prefer LL for reachability to
other LL addresses on link. This will cause neighbor solicitations
to be sent using LL to be sent and result in that address being
resolved on servers. In absence of RA.routerAddress config OTOH
we prefer, the well known address. This allows for hitless
upgrade, downgrade

Differential Revision: D39116389

fbshipit-source-id: 0a7dbd6ed1a0b7922030d9ade8d377e3584d4ce4