Make parent scheduler node attribute optional, add default fallback
Summary:
Indus ASIC does not support this attribute. I am working
with the vendor to implement this but its not there currently.
AFAIK, we don't really use this attribute for anything beyond
get, so functionality wise this should be no-op.

Interestingly per the sai spec, this attribute is mandatory on
create. So the vendor sai-sdk should have set and returned a
value for all the pre created queues. Unfortunately they don't

Reviewed By: zechengh09

Differential Revision: D39162048

fbshipit-source-id: e568e7e6799d36c3a5293e8ae2803f745a242e34