Added interfaceType attribute for PHY port creation
Summary:
For Sandia 100G port configuration, we need to pass the interface type to Port Create function. This diff adds the passing of interface type for Sandia (for Credo platform it will continue passing std::nullopt as earlier).
For all existing platforms, interface type is derived from medium and speed value. In this platform also we continue to do the same.

Reviewed By: harshitgulati18

Differential Revision: D38495459

fbshipit-source-id: 1b37c6f40f04b7782fac3d3c128c09c05f7c6efd