Add INTERNAL scheduling type enum to cater to fabric port queues
Summary:
For fabric port queues the NOS doesn't really have any
knobs exposed for programming, introspection. So add a internal
scheduling type enum to cater to that case. I considered
alternatively to make PortQueue.scheduling optional
https://github.com/facebook/fboss/blob/main/fboss/agent/switch_config.thrift#L699

However, that field is mark required and changing to
optional is not backwards compatible -
https://www.internalfb.com/intern/wiki/Thrift/Thrift_Guide/Best_Practices/

Reviewed By: nivinl

Differential Revision: D39358020

fbshipit-source-id: 656b281ec3e54c9121a2f20f965f57a1671ac24c