rackmon2: Add debug trace capability
Summary:
Add support to (default off) optionally print a trace of
logs received/transmitted message over any of the supported
interfaces.

This is implemented as a interface configuration so we can use it to
debug specific issues on an interface rather than globally, and also
not require special debug builds.

Imported from: D38556571

Reviewed By: tao-ren

Differential Revision: D39297062

fbshipit-source-id: 806e64b5c83e5ab0b291171d5c3626e8f1d8b430