rackmon2: Fix rare unit test failure.
Summary:
Yeah, dont write unit tests which depend on timing
or system load. But thats the only way to test this
one unit after which all things depending on it
can be tested using mocks.

So, make the expectation between(2,3) instead of exactly
2. This allows us wiggle room when the test took just
a bit longer to stop and we had an additional call.

Imported from: D39297407

Reviewed By: somasun

Differential Revision: D39323246

fbshipit-source-id: cc9e61a5306f14c7b77b5f950b39150be918491e