Temporary hack to get initial sync working
Summary:
Turns out, our initial sync never worked. We try to publish the current state and then start(). Except that until we call start(), the syncer is not ready so all calls to publish just no-ops.

This entire code will be completely replaced very soon in D38564164. But until that lands. Just mark it as ready sooner. It's possibly that changes got published in this gap and we will end up overwriting the new state with the old. But this currently doesn't work at all. So at least we're not making things worse and a fulls solution is coming soon.

Reviewed By: peygar

Differential Revision: D39151264

fbshipit-source-id: 3f44e0637ddf9bbd321a197b4150463829caf141