Add CPU to port types, since it is a type of port
Summary:
In certain contexts we would like like to handle
all port types - nif, fabric and cpu uniformly. Viz when
getting supported stream types today we have a boolean
check on port type being cpu or not. Now we need to
add another param to convey whether port type is fabric
or not. Having a single enum makes this uniform

Reviewed By: zechengh09

Differential Revision: D39521678

fbshipit-source-id: 58544dc71e0c9b91275fce8566ccbe57e6197e61