Gate control plane config application on CPU port feature
Summary:
As titled. Also assert that control plane config
is null in case CPU port is not supported. We need to get
on ASIC support as in its absence we will try to query
the default cpu port stream types, num queues etc. This
will throw on ASICs where there are no CPU ports

Reviewed By: zechengh09

Differential Revision: D39547294

fbshipit-source-id: b92b46c0e40e03620c009e74e37747805af8e1fe