Refactor the logic to detect an active qsfp_service
Summary:
This diff is part of an effort to get all applicable wedge_qsfp_util commands to work with/without qsfp_service. Currently, there are quite a few commands that access the HW directly and thus, interfere with an active qsfp_service. These commands will need to detect an active qsfp_service and route their requests to qsfp_service.

This diff refactors the logic to detect an active qsfp_service into the QsfpServiceDetector singleton to be shared by all wedge_qsfp_util commands.

Reviewed By: harshitgulati18

Differential Revision: D39587506

fbshipit-source-id: 958baeed48bae4e93b60e19a58ed39d3d079d96d