Update comment on interface vlan config to reflect non VLAN rifs
Summary:
We are working on adding support for port based RIFs.
Initially, I just changed this field to be optional D39624919
unfortunately there are a lot of places where this vlanID
field is used. While its possible to fix all references in
compiled language code (since the type change) and all code
that is covered via tests. Unfortunately there is bound to
be code that falls out of either of these 2 buckets. So for
now only updating the comment to reflect the upcoming change.
Post the feature (port RIFs) we will need to make changes in
code that always assume interface types of VLANs, but that
breakage would be restricted to DSF, which is a new product
which we can fix as we develop.

Reviewed By: shri-khare

Differential Revision: D39662618

fbshipit-source-id: 7f46274ea29b1ffd5d711a186f024de3b532f81b