Fsdb syncer library based on one cow storage mgr
Summary:
Rewriting syncer library to better handle initial sync. The design here is much simpler, the syncer will have a managed cow storage that different components will queue updates on. No publishing will happen until start() is called. The binary is in charge of determining when is a good time to start publishing. Once start() is called, we'll do a full sync of our internal state then flip on delta publishing.

Similarly during times of disconnection, we're going to still apply updates to the internal state but skip the publishing until we're reconnected.

This library will interop particularly nicely when the service is representing the state of it's different components as CowStorage already. The updateState routine will be reduced to a trivial copy of the component's root to the service's root struct.

Since this is essentially a complete rewrite, I opted to name this as FsdbSyncManager2 for the time being, to make the review easier and so I can split mka refactor to a separate diff.

Reviewed By: jasmeetbagga

Differential Revision: D39074711

fbshipit-source-id: 78d2fb05d7e2111290235f5421bc08e84c2c6104