Serialize add and removal of stats entries in BcmStatsUpdater
Summary: The stats updater used 2 queues 1) to be removed stats 2) to be added stats and it periodically checks both queues and processes removal first and then add queue. So if there is a remove + add + remove, we will end up doing a remove and then an add. This does not work for fast updating stats such as TE stats. So move add and remove to a single queue and process them serially

Reviewed By: rskumaran-fb

Differential Revision: D39180131

fbshipit-source-id: 4027f33780b7107383964bee18d650637d35f4c9