Refactor the logic to get and store the transceiver bus
Summary:
This diff is part of an effort to get all applicable wedge_qsfp_util commands to work with/without qsfp_service. Currently, there are quite a few commands that access the HW directly and thus, interfere with an active qsfp_service. These commands will need to detect an active qsfp_service and route their requests to qsfp_service.

This diff moves the logic to get the transceiver bus from qsfp_util_main to the QsfpUtilContainer singleton and stores the unique_ptr to this bus there. All wedge_qsfp_util commands can then get this pointer as needed.

Reviewed By: harshitgulati18

Differential Revision: D39589357

fbshipit-source-id: 6d96fdaa7b24b3fbdc699b5f9d8e3532814e791a