Create readRegister() function that works with and without qsfp_service
Summary:
This diff is part of an effort to get all applicable wedge_qsfp_util commands to work with/without qsfp_service. Currently, there are quite a few commands that access the HW directly and thus, interfere with an active qsfp_service. These commands will need to detect an active qsfp_service and route their requests to qsfp_service.

This diff creates a new readRegister() function that works with and without the --direct-i2c option so that these commands can just call it. The next diff will add a new writeRegister() function.

Reviewed By: harshitgulati18

Differential Revision: D39664196

fbshipit-source-id: 07f633b076c48db4b311d3063f9e822b55d2bd57