Fix the set/clear low power commands to work with and without active qsfp_service
Summary:
This diff is part of an effort to get all applicable wedge_qsfp_util commands to work with/without qsfp_service. Currently, there are quite a few commands that access the HW directly and thus, interfere with an active qsfp_service. These commands will need to detect an active qsfp_service and route their requests to qsfp_service.

The set/clear low power commands fall into this category in which they always set/clear low power of optics directly regardless of whether or not qsfp_service is running. This diff fixes them.

Reviewed By: harshitgulati18

Differential Revision: D39669409

fbshipit-source-id: ff47825144b23a7464b8004f187590b8f6c3ebca