Added extended Fec for Phy port creation from SAI 1.10.0
Summary:
The Extended FEC mode is required by the Sandia PHY creation. For 400G port, it has to be RS544 on both system and line side. For 10G port it needs to be RS544 on host side and RS528 on the line side.
Based on the Extended FEC mode, the PHY selects correct operation mode (they call is MZDP100UP2 etc). So this required at the time f port create.
Adding this attribute for port create.

Reviewed By: harshitgulati18

Differential Revision: D39137481

fbshipit-source-id: 3b3f1ba5ce1c6d7aa0c6e4ccd5f81f3f351f90a2