Fix timing bugs in BidirectionalPacketStreamTest
Summary:
Contbuild has been failing:
https://www.internalfb.com/svc/conveyors/fboss/wedge_agent/bundles/2243?node_name=build&run_id=110714792
https://www.internalfb.com/intern/sandcastle/job/36028797793984824/insights

This seems timing related. We need to wait for the stream to actually get connected before proceeding with the test.

Also, reset the pointer to local packet acceptor before we go out of scope and gets deleted.

Finally, made some improvements to debug messages in the area.

Reviewed By: rajank7

Differential Revision: D39718020

fbshipit-source-id: 843dbf02f3df6ce17cf67b3be49e80491ee9d45a