Get neighbor tests ready of VOQ switches
Summary:
For Voq switches, since we use port based
RIFs we need to use onePortPerInterface config. Make
it so in Neighbor tests when not running in LAG mode.

For LAGs in VOQ switch, we will need to get more sophisticated
and associate system ports with LAGs and not physical ports.
But lets handle the non LAG case first, LAGs are not in
the day 1 use case

Reviewed By: nivinl

Differential Revision: D39771389

fbshipit-source-id: 65b479daae5692fd85beccc02796bb51528bc541