Add/remove neighbor test for Voq switch
Summary:
As titled. Note that ultimately we do want to
use HwNeighborTests - https://fburl.com/code/npipmk3e.
However we are still working on getting data plane tests
working on Voq switches, so need to hold off on that for
a bit.
Meanwhile, try to exercise nbr programming with a simple
add, remove neighbor test.

Reviewed By: nivinl

Differential Revision: D39782437

fbshipit-source-id: 6790bfbb579f633159e294a32732906600fd20c8